Supports standard R/C car configuration (1.3)
This release adds support for the standard R/C car configuration using
one channel for steering and another channel for speed (ESC). A checkbox
in the app prefs "Mixing activated" is employed to switch between the
two modes. In order to maintain backwards compatibility the app comes
with mixing activated ("tank mode with steering by controlling the speed
of the left or right motor").