Supports fail safe function (1.4)
This release fully supports a fail safe function which is activated in
the SSC Settings. If you want to maintain the current time-out value of
your PiKoder then start off with reading the EEPROM and then store the
value back to the PiKoder's memory. Please note that the time out value
is specified in 1/10th of a second.